+++ src/main/java/io/github/apricotfarmer11/mods/tubion/TubionMod.java
@@ -5,12 +5,12 @@ import io.github.apricotfarmer11.mods.tubion.core.tubnet.TubnetCore;
 import io.github.apricotfarmer11.mods.tubion.config.TubionConfig;
 import io.github.apricotfarmer11.mods.tubion.config.TubionConfigManager;
 import io.github.apricotfarmer11.mods.tubion.core.tubnet.event.TubnetConnectionEvents;
-import io.github.apricotfarmer11.mods.tubion.event.ChatMessageEvent;
 import io.github.apricotfarmer11.mods.tubion.feature.AutoGGL;
 import io.github.apricotfarmer11.mods.tubion.feature.ChatUtils;
 import io.github.apricotfarmer11.mods.tubion.feature.discord.SDKInitializer;
 import io.github.apricotfarmer11.mods.tubion.feature.discord.TubnetDiscordIntegration;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.loader.api.FabricLoader;
 //#if MC>=11902
 //$$ import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
@@ -19,6 +19,8 @@ import net.fabricmc.fabric.api.client.command.v1.ClientCommandManager;
 //#endif
 
 import java.io.IOException;
+import java.net.InetSocketAddress;
+import java.net.SocketAddress;
 import java.util.HashMap;
 import java.util.UUID;
 
@@ -30,6 +32,7 @@ public class TubionMod implements ClientModInitializer {
     public static TubnetDiscordIntegration discordIntegration;
     public static TubionCommandManager commandManager;
     private static TubionMod instance;
+    public static boolean deferReload = false;
     public void onInitializeClient() {
         instance = this;
         try {
@@ -46,9 +49,9 @@ public class TubionMod implements ClientModInitializer {
         commandManager.loadCommands();
         //#endif
 
-        // Load chat utilities
-        ChatMessageEvent.EVENT.register(new ChatUtils()::onChat);
+        // Load ui utilities
         new AutoGGL();
+        ChatUtils.register();
         TubnetConnectionEvents.CONNECT.register(() -> {
             this.initDiscord();
         });
@@ -56,9 +59,16 @@ public class TubionMod implements ClientModInitializer {
             discordIntegration.destroy();
             discordIntegration = null;
         });
+        ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> {
+            if (handler.getConnection().getAddress() instanceof InetSocketAddress) {
+                if (((InetSocketAddress) handler.getConnection().getAddress()).getHostName().toLowerCase().matches("^((.*)\\.)?tubnet\\.gg\\.?$")) {
+                    deferReload = true;
+                }
+            }
+        });
     }
     public void initDiscord() {
-        if (discordIntegration != null && discordIntegration.discordGameSdk != null && discordIntegration.discordGameSdk.isOpen()) {
+        if (discordIntegration != null) {
             discordIntegration.destroy();
             discordIntegration = null;
         }