+++ src/main/java/io/github/apricotfarmer11/mods/tubion/core/tubient/model/social/TubnetParty.java
@@ -1,0 +1,10 @@
+package io.github.apricotfarmer11.mods.tubion.core.tubient.model.social;
+
+/**
+ * Class for party-related stuff
+ */
+public class TubnetParty {
+    private TubnetPlayer partyOwner;
+
+
+}