+++ src/main/java/io/github/apricotfarmer11/mods/tubion/core/tubient/model/types/PlayerEdition.java
@@ -1,0 +1,6 @@
+package io.github.apricotfarmer11.mods.tubion.core.tubient.model.types;
+
+public enum PlayerEdition {
+    JAVA,
+    BEDROCK
+}