+++ src/main/java/io/github/apricotfarmer11/mods/tubion/event/api/types/EventType.java
@@ -1,0 +1,15 @@
+package io.github.apricotfarmer11.mods.tubion.event.api.types;
+
+public class EventType {
+
+    /**
+     * Used to define the type of typed event.
+     */
+    public static final byte
+            PRE = 0,
+            ON = 1,
+            POST = 2,
+            SEND = 3,
+            RECEIVE = 4;
+
+}