+++ src/main/java/io/github/apricotfarmer11/mods/tubion/feature/AutoGGL.java
@@ -4,8 +4,12 @@ import io.github.apricotfarmer11.mods.tubion.TubionMod;
 import io.github.apricotfarmer11.mods.tubion.core.helper.PlayerHelper;
 import io.github.apricotfarmer11.mods.tubion.core.tubnet.TubnetCore;
 import io.github.apricotfarmer11.mods.tubion.core.tubnet.game.GameMode;
-import io.github.apricotfarmer11.mods.tubion.event.ChatMessageEvent;
 import io.github.apricotfarmer11.mods.tubion.event.GameHudEvents;
+import io.github.apricotfarmer11.mods.tubion.event.api.EventManager;
+import io.github.apricotfarmer11.mods.tubion.event.api.EventTarget;
+import io.github.apricotfarmer11.mods.tubion.event.api.events.Event;
+import io.github.apricotfarmer11.mods.tubion.event.ui.ChatMessageEvent;
+import io.github.apricotfarmer11.mods.tubion.event.ui.TitleModifyEvent;
 import net.minecraft.util.ActionResult;
 
 import java.util.Arrays;
@@ -18,19 +22,22 @@ public class AutoGGL {
     };
 
     public AutoGGL() {
-        GameHudEvents.TITLE_SET.register((text) -> {
-            if (Arrays.stream(COMPETITIVE_GAMES).anyMatch((a) -> a == TubnetCore.getInstance().getGameMode())) {
-                // GLHF
-                if (TubionMod.getConfig().autoGf && text.getString().contains("go!")) {
-                    PlayerHelper.sendChatMessage("glhf");
-                }
-            }
-        });
-        ChatMessageEvent.EVENT.register(message -> {
-            if (Arrays.stream(COMPETITIVE_GAMES).anyMatch(a -> a == TubnetCore.getInstance().getGameMode()) && TubionMod.getConfig().autoGg && (message.getString().contains("Game finished") || message.getString().contains("Game over") || message.getString().contains("Your team took"))) {
-                PlayerHelper.sendChatMessage("gg");
-            }
-            return ActionResult.PASS;
-        });
+        EventManager.register(this);
+    }
+
+    @EventTarget
+    public void onChat(ChatMessageEvent ev) {
+        String message = ev.getMessage().getString();
+        if (Arrays.stream(COMPETITIVE_GAMES).anyMatch(a -> a == TubnetCore.getInstance().getGameMode()) && TubionMod.getConfig().autoGg && message.matches("(Game finished|Game over|Your team took)")) {
+            PlayerHelper.sendChatMessage("gg");
+        }
+    }
+
+    @EventTarget
+    public void onTitleUpdate(TitleModifyEvent ev) {
+        String title = ev.getMessage().getString();
+        if (title.equals("go!") && Arrays.stream(COMPETITIVE_GAMES).anyMatch(a -> a == TubnetCore.getInstance().getGameMode()) && TubionMod.getConfig().autoGf) {
+            PlayerHelper.sendChatMessage("glhf");
+        }
     }
 }