+++ src/main/java/io/github/apricotfarmer11/mods/tubion/mixin/gui/InGameHudMixin.java
@@ -1,6 +1,8 @@
 package io.github.apricotfarmer11.mods.tubion.mixin.gui;
 
 import io.github.apricotfarmer11.mods.tubion.event.GameHudEvents;
+import io.github.apricotfarmer11.mods.tubion.event.api.EventManager;
+import io.github.apricotfarmer11.mods.tubion.event.ui.TitleModifyEvent;
 import net.minecraft.client.gui.hud.InGameHud;
 import net.minecraft.text.Text;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,6 +14,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public abstract class InGameHudMixin {
     @Inject(at = @At("HEAD"), method = "setTitle")
     public void setTitle(Text title, CallbackInfo ci) {
-        GameHudEvents.TITLE_SET.invoker().onTitleSet(title);
+        EventManager.call(new TitleModifyEvent(title));
     }
 }