+++ CHANGELOG.md
@@ -6,3 +6,6 @@ Heavy technical changes. Don't expect new features in v0.2.0. v0.3.0 will come o
 
 ### New Features
 * New Feature: "Hide Tips"
+* New panorama, created by TubNet themselves!
+* You now can use the Tubion discord bot in the [official tubion server](https://dsc.gg/tubion). This bot allows you to check player info, and more soon!
+* Refactored event system